RemoveUser from construction done

+++ src/main/kotlin/isel/pt/ps/projeto/services/ConstructionsService.kt
@@ -115,8 +115,13 @@ class ConstructionsService(
     fun removeConstructionUser(authId: Int, oid: Int, uid: Int): RemoveUserFromConstructionResult {
         val role = constructionsRepository.getUserRoleFromConstruction(authId, oid)
             ?: return failure(ConstructionUserError.NoPermission)
+
+        val userToBeRemoved = constructionsRepository.getUserRoleFromConstruction(uid, oid)
+            ?: return failure(ConstructionUserError.UserNotFound)
+
+        if (userToBeRemoved.role == "admin") return failure(ConstructionUserError.NoPermission)
         if (role.role != "admin") return failure(ConstructionUserError.NoPermission)
-        constructionsRepository.getConstructionUser(oid, uid) ?: return failure(ConstructionUserError.UserNotFound)
+        
         return success(constructionsRepository.removeConstructionUser(oid, uid))
     }
